chore(precompiles): feature gate reth dependencies (#395)
Similar to https://github.com/tempoxyz/tempo/pull/384 in motivation, we
want to slim `tempo-precompiles` to a degree where we can integrate it
into Foundry without pulling in unnecessary dependencies. Foundry does
not (as of now) depend on any Reth crates.

This feature gates the `TIPFeeStateProviderExt` trait on a new feature,
`reth`.